Refactor NoSuchEventTypeException
