+++ src/main/java/org/zalando/nakadi/exceptions/runtime/NoEventTypeException.java
@@ -1,7 +1,5 @@
 package org.zalando.nakadi.exceptions.runtime;
 
-import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
-
 public class NoEventTypeException extends MyNakadiRuntimeException1 {
 
     public NoEventTypeException(final String msg) {
