+++ src/main/java/org/zalando/nakadi/exceptions/runtime/NoSuchEventTypeException.java
@@ -1,0 +1,12 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class NoSuchEventTypeException extends MyNakadiRuntimeException1 {
+    public NoSuchEventTypeException(final String message) {
+        super(message);
+    }
+
+    public NoSuchEventTypeException(final String msg, final Exception cause) {
+        super(msg, cause);
+    }
+
+}
