+++ src/main/java/org/zalando/nakadi/service/NakadiCursorComparator.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
-import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
 import org.zalando.nakadi.repository.db.EventTypeCache;
 
 import java.util.Comparator;
