+++ src/main/java/org/zalando/nakadi/service/converter/VersionedConverter.java
@@ -3,7 +3,7 @@ package org.zalando.nakadi.service.converter;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.InvalidCursorException;
-import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.service.CursorConverter;
 import org.zalando.nakadi.view.Cursor;
