+++ src/main/java/org/zalando/nakadi/service/subscription/SubscriptionService.java
@@ -24,7 +24,7 @@ import org.zalando.nakadi.domain.SubscriptionEventTypeStats;
 import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.InvalidCursorException;
-import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
+import org.zalando.nakadi.exceptions.runtime.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.NoSuchSubscriptionException;
 import org.zalando.nakadi.exceptions.Try;
 import org.zalando.nakadi.exceptions.runtime.DbWriteOperationsBlockedException;
@@ -66,6 +66,8 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
+
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 
 @Component
 public class SubscriptionService {
@@ -187,7 +189,7 @@ public class SubscriptionService {
             return Result.ok(subscription);
         } catch (final NoSuchSubscriptionException e) {
             LOG.debug("Failed to find subscription: {}", subscriptionId);
-            return Result.problem(Problem.valueOf(Response.Status.NOT_FOUND, e.getMessage()));
+            return Result.problem(Problem.valueOf(NOT_FOUND, e.getMessage()));
         } catch (final ServiceTemporarilyUnavailableException e) {
             LOG.error("Error occurred when trying to get subscription: {}", subscriptionId, e);
             return Result.problem(Problem.valueOf(Response.Status.SERVICE_UNAVAILABLE, e.getMessage()));
@@ -214,11 +216,14 @@ public class SubscriptionService {
             return Result.ok();
         } catch (final NoSuchSubscriptionException e) {
             LOG.debug("Failed to find subscription: {}", subscriptionId, e);
-            return Result.problem(Problem.valueOf(Response.Status.NOT_FOUND, e.getMessage()));
+            return Result.problem(Problem.valueOf(NOT_FOUND, e.getMessage()));
         } catch (final ServiceTemporarilyUnavailableException e) {
             LOG.error("Error occurred when trying to delete subscription: {}", subscriptionId, e);
             return Result.problem(Problem.valueOf(Response.Status.SERVICE_UNAVAILABLE, e.getMessage()));
-        } catch (final NoSuchEventTypeException | InternalNakadiException e) {
+        } catch (final NoSuchEventTypeException e) {
+            LOG.error("Event Type not found", e);
+            return Result.problem(Problem.valueOf(NOT_FOUND, e.getMessage()));
+        } catch (final InternalNakadiException e) {
             LOG.error("Exception can not occur", e);
             return Result.problem(e.asProblem());
         }
